Restructuring `MATCH` and `Gremlin` lowering into respective directories (#76)
* Moved lowering steps for optional traversal into new file.

* Restructured lowering code.

* Removed BetweenClause from ir_lowering_match

* removed unused import

* removed unused import

* Renamed files

* Adding missing files
