+++ graphql_compiler/tests/test_emit_output.py
@@ -9,7 +9,7 @@ from ..compiler.expressions import (BinaryComposition, ContextField, LocalField,
                                     OutputContextField, TernaryConditional, Variable)
 from ..compiler.helpers import Location
 from ..compiler.ir_lowering_common import OutputContextVertex
-from ..compiler.ir_lowering_match import CompoundMatchQuery
+from ..compiler.ir_lowering_match.optional_traversal import CompoundMatchQuery
 from ..compiler.match_query import convert_to_match_query
 from ..schema import GraphQLDateTime
 from .test_helpers import compare_gremlin, compare_match
